requirements.txt
altair
altair_saver
debugpy==1.5.1
disjoint-set==0.7.3
duckdb==0.5.1
fastcluster==1.2.4
ipykernel==6.9.1
jsonschema==4.16.0 # https://github.com/altair-viz/altair/issues/2767
matplotlib==3.5.1
notebook
numpy==1.22.0
opencv-python==4.5.5.62
pandas==1.3.5
Pillow==9.0.0
pyarrow==8.0.0
pytorchvideo==0.1.5
scikit-learn==1.0.2
scikit-learn-extra==0.2.0
scipy==1.7.3
seaborn==0.12.2
streamlit
